refactor: rename svelte tabs component to TabbedContent
