+++ packages/upswyng-server/src/components/TabbedContent.svelte
